Added order checkout and placeing order

+++ foody/common/src/main/java/com/foody/common/model/response/restaurant/ItemResponse.java
@@ -2,6 +2,7 @@ package com.foody.common.model.response.restaurant;
 
 import com.foody.common.model.request.restaurant.ItemRequest;
 import com.foody.common.model.request.restaurant.ItemType;
+import com.foody.common.model.response.order.Price;
 import lombok.*;
 
 import java.math.BigDecimal;
@@ -16,13 +17,13 @@ public class ItemResponse {
     private Integer restaurantId;
     private String itemId;
     private String unitType;
-    private Double quantity;
+    private double quantity;
     private ItemType itemType;
     private String itemName;
     private String itemDescription;
     private String itemCategoryCode;
     private String itemCategoryDescription;
-    private BigDecimal price;
+    private double price;
     private boolean availability;
 
     //Image of items are stored in external drive naming restaurantId_itemId and accessing via URL
